Change exports and imports of format functions
