+++ src/components/CountryCard.jsx
@@ -8,67 +8,7 @@ import { Link } from "react-router-dom";
 import { addFavourite, removeFavourite, updateAlertProps } from "../store/favouritesSlice";
 import { useDispatch, useSelector } from "react-redux";
 import { addFavouriteToFirebase, auth } from "../auth/firebase";
-
-// Return output separates items with comma
-const formatLanguages = (languagesObject) => {
-  if (!languagesObject) {
-    return []
-  }
-
-  const languages = Object.values(languagesObject);
-  if (languages.length === 0) {
-    return [];
-  }
-
-  const formattedLanguages = [];
-  if (languages.length === 1) {
-    return languages;
-  }
-  for (let i = 0; i < languages.length; i++) {
-    if (languages.length - 1 === i) {
-      formattedLanguages.push(languages[i]);
-    }
-    else if (languages.length - 2 === i) {
-      formattedLanguages.push(languages[i] + " & ")
-    }
-    else {
-      formattedLanguages.push(languages[i] + ", ")
-    }
-  }
-  return formattedLanguages
-}
-
-// Return output separates items with comma
-const formatCurrencies = (currencyObject) => {
-  if (!currencyObject) {
-    return [];
-  }
-
-  const currencies = Object.values(currencyObject);
-
-  if (currencies.length === 0) {
-    return [];
-  }
-
-  if (currencies.length === 1) {
-    return currencies[0].name;
-  }
-  else {
-    const formattedCurrencies = [];
-    for (let i = 0; i < currencies.length; i++) {
-      if (i === currencies.length - 1) {
-        formattedCurrencies.push(currencies[i].name);
-      }
-      else if (i === currencies.length - 2) {
-        formattedCurrencies.push(currencies[i].name + " & ")
-      }
-      else {
-        formattedCurrencies.push(currencies[i].name + ", ")
-      }
-    }
-    return formattedCurrencies;
-  }
-}
+import { formatCurrencies, formatLanguages } from "../utils/helperFunctions";
 
 const CountryCard = ({ country }) => {
   // Flag is in 2∶1 in aspect ratio. Use this value to adjust the scale the flag size correctly.
@@ -151,4 +91,3 @@ const CountryCard = ({ country }) => {
 }
 
 export default CountryCard
-export { formatCurrencies, formatLanguages }
